数据字典

+++ pkg/db/ent/dictionarydetail.go
@@ -18,13 +18,13 @@ type DictionaryDetail struct {
 	config `json:"-"`
 	// ID of the ent.
 	// primary key
-	ID uint64 `json:"id,omitempty"`
+	ID int64 `json:"id,omitempty"`
 	// created time
 	CreatedAt time.Time `json:"created_at,omitempty"`
 	// last update time
 	UpdatedAt time.Time `json:"updated_at,omitempty"`
 	// status 1 normal 0 ban | 状态 1 正常 0 禁用
-	Status uint8 `json:"status,omitempty"`
+	Status int8 `json:"status,omitempty"`
 	// the title shown in the ui | 展示名称 （建议配合i18n）
 	Title string `json:"title,omitempty"`
 	// key | 键
@@ -32,7 +32,7 @@ type DictionaryDetail struct {
 	// value | 值
 	Value string `json:"value,omitempty"`
 	// Dictionary ID | 字典ID
-	DictionaryID uint64 `json:"dictionary_id,omitempty"`
+	DictionaryID int64 `json:"dictionary_id,omitempty"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the DictionaryDetailQuery when eager-loading is set.
 	Edges        DictionaryDetailEdges `json:"edges"`
@@ -92,7 +92,7 @@ func (dd *DictionaryDetail) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			dd.ID = uint64(value.Int64)
+			dd.ID = int64(value.Int64)
 		case dictionarydetail.FieldCreatedAt:
 			if value, ok := values[i].(*sql.NullTime); !ok {
 				return fmt.Errorf("unexpected type %T for field created_at", values[i])
@@ -109,7 +109,7 @@ func (dd *DictionaryDetail) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field status", values[i])
 			} else if value.Valid {
-				dd.Status = uint8(value.Int64)
+				dd.Status = int8(value.Int64)
 			}
 		case dictionarydetail.FieldTitle:
 			if value, ok := values[i].(*sql.NullString); !ok {
@@ -133,7 +133,7 @@ func (dd *DictionaryDetail) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field dictionary_id", values[i])
 			} else if value.Valid {
-				dd.DictionaryID = uint64(value.Int64)
+				dd.DictionaryID = value.Int64
 			}
 		default:
 			dd.selectValues.Set(columns[i], values[i])
