+++ pkg/db/ent/user.go
@@ -18,13 +18,13 @@ type User struct {
 	config `json:"-"`
 	// ID of the ent.
 	// primary key
-	ID uint64 `json:"id,omitempty"`
+	ID int64 `json:"id,omitempty"`
 	// created time
 	CreatedAt time.Time `json:"created_at,omitempty"`
 	// last update time
 	UpdatedAt time.Time `json:"updated_at,omitempty"`
 	// status 1 normal 0 ban | 状态 1 正常 0 禁用
-	Status uint8 `json:"status,omitempty"`
+	Status int8 `json:"status,omitempty"`
 	// user's login name | 登录名
 	Username string `json:"username,omitempty"`
 	// password | 密码
@@ -38,7 +38,7 @@ type User struct {
 	// active color of template | 当前激活的颜色设定
 	ActiveColor string `json:"active_color,omitempty"`
 	// role id | 角色ID
-	RoleID uint64 `json:"role_id,omitempty"`
+	RoleID int64 `json:"role_id,omitempty"`
 	// mobile number | 手机号
 	Mobile string `json:"mobile,omitempty"`
 	// email | 邮箱号
@@ -106,7 +106,7 @@ func (u *User) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			u.ID = uint64(value.Int64)
+			u.ID = int64(value.Int64)
 		case user.FieldCreatedAt:
 			if value, ok := values[i].(*sql.NullTime); !ok {
 				return fmt.Errorf("unexpected type %T for field created_at", values[i])
@@ -123,7 +123,7 @@ func (u *User) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field status", values[i])
 			} else if value.Valid {
-				u.Status = uint8(value.Int64)
+				u.Status = int8(value.Int64)
 			}
 		case user.FieldUsername:
 			if value, ok := values[i].(*sql.NullString); !ok {
@@ -165,7 +165,7 @@ func (u *User) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field role_id", values[i])
 			} else if value.Valid {
-				u.RoleID = uint64(value.Int64)
+				u.RoleID = value.Int64
 			}
 		case user.FieldMobile:
 			if value, ok := values[i].(*sql.NullString); !ok {
